AXON-1372: add assignee column
