+++ src/webviews/components/issue/LinkedIssues.tsx
@@ -1,14 +1,11 @@
 import Button from '@atlaskit/button';
-import DropdownMenu, { DropdownItem, DropdownItemGroup } from '@atlaskit/dropdown-menu';
-import ChevronDownIcon from '@atlaskit/icon/core/chevron-down';
-import Lozenge from '@atlaskit/lozenge';
 import TableTree from '@atlaskit/table-tree';
 import Tooltip from '@atlaskit/tooltip';
 import { IssueLinkIssue, MinimalIssueLink, MinimalIssueOrKeyAndSite } from '@atlassianlabs/jira-pi-common-models';
 import * as React from 'react';
 
 import { DetailedSiteInfo } from '../../../atlclients/authInfo';
-import { colorToLozengeAppearanceMap } from '../colors';
+import { AssigneeColumn, Priority, StatusColumn, Summary } from './IssueColumns';
 
 type LinkedIssuesProps = {
     issuelinks: MinimalIssueLink<DetailedSiteInfo>[];
@@ -53,90 +50,6 @@ const IssueKey = (data: ItemData) => {
     );
 };
 
-const Summary = (data: ItemData) => <p style={{ display: 'inline' }}>{data.issue.summary}</p>;
-const Priority = (data: ItemData) => {
-    if (!data.issue.priority || !data.issue.priority.name || !data.issue.priority.iconUrl) {
-        return <React.Fragment />;
-    }
-
-    return (
-        <div style={{ width: '16px', height: '16px' }}>
-            <Tooltip content={data.issue.priority.name}>
-                <img src={data.issue.priority.iconUrl} alt={data.issue.priority.name} />
-            </Tooltip>
-        </div>
-    );
-};
-
-const StatusColumn = (data: ItemData) => {
-    if (!data.issue.status || !data.issue.status.statusCategory) {
-        return <React.Fragment />;
-    }
-
-    if (data.onStatusChange) {
-        const currentStatus = data.issue.status.name;
-        const currentLozColor = colorToLozengeAppearanceMap[data.issue.status.statusCategory.colorName] || 'default';
-
-        const transitions = (data.issue as any).transitions || [];
-
-        const validTransitions = transitions.filter((transition: any) => transition.to.id !== data.issue.status.id);
-
-        if (validTransitions.length === 0) {
-            return <Lozenge appearance={currentLozColor}>{currentStatus}</Lozenge>;
-        }
-
-        return (
-            <div style={{ width: '150px', fontSize: '12px', minWidth: '140px' }}>
-                <DropdownMenu
-                    trigger={({ triggerRef, ...props }) => (
-                        <Button
-                            {...props}
-                            ref={triggerRef}
-                            appearance="subtle"
-                            style={{
-                                padding: '4px 6px',
-                                minHeight: '32px',
-                            }}
-                            iconAfter={<ChevronDownIcon label="" size="small" />}
-                        >
-                            <Lozenge appearance={currentLozColor}>{currentStatus}</Lozenge>
-                        </Button>
-                    )}
-                    placement="bottom-start"
-                >
-                    <DropdownItemGroup>
-                        {validTransitions.map((transition: any) => {
-                            const lozColor =
-                                colorToLozengeAppearanceMap[transition.to.statusCategory.colorName] || 'default';
-
-                            return (
-                                <DropdownItem
-                                    key={transition.id}
-                                    onClick={() => {
-                                        if (data.onStatusChange) {
-                                            data.onStatusChange(data.issue.key, transition.to.name);
-                                        }
-                                    }}
-                                >
-                                    <Lozenge appearance={lozColor}>{transition.to.name}</Lozenge>
-                                </DropdownItem>
-                            );
-                        })}
-                    </DropdownItemGroup>
-                </DropdownMenu>
-            </div>
-        );
-    }
-
-    const lozColor: string = colorToLozengeAppearanceMap[data.issue.status.statusCategory.colorName];
-    return <Lozenge appearance={lozColor}>{data.issue.status.name}</Lozenge>;
-};
-// const Delete = (data: ItemData) => {
-//     return (<div className='ac-delete' onClick={() => data.onDelete(data.issue)}>
-//         <TrashIcon label='trash' />
-//     </div>);
-// };
-
 export const LinkedIssues: React.FunctionComponent<LinkedIssuesProps> = ({
     issuelinks,
     onIssueClick,
@@ -145,8 +58,8 @@ export const LinkedIssues: React.FunctionComponent<LinkedIssuesProps> = ({
 }) => {
     return (
         <TableTree
-            columns={[IssueKey, Summary, Priority, StatusColumn]}
-            columnWidths={['150px', '100%', '20px', '150px']}
+            columns={[IssueKey, Summary, Priority, AssigneeColumn, StatusColumn]}
+            columnWidths={['150px', '100px', '20px', '150px', '150px']}
             items={issuelinks.map((issuelink) => {
                 return {
                     id: issuelink.id,
